Python 2 to 3 fixes for src/utils/
* The majority of the code in src/utils/ was only fully compatible with
  Python 2, including depending on parts of the standard library which
  have been drastically changed between Python 2 and Python 3.
* Lots of bugs have been occurring and being mostly ignored (usually
  only recorded at the INFO level in the message.log).
* This change consists of running 2to3 on all the files and making
  versions for both Python 2 and 3 where appropriate.  In each case the
  replacement file checks the major version number of Python and imports
  the relevant code for that version.
* This has been done file by file in case a plugin or other resource
  calls one of those components explicitly (e.g. as utils.web), rather
  than putting all the version checking in __init__.py.
* It is possible that there may still be some bugs related to these
  older modules, but they will be more readily identified without having
  to decipher what is the result of Python 2 and 3 conflicts.
* Of likely concern, however, is the use of submodule names which clash
  with builtins (e.g. file.py), the python submodule (python.py) and
  instances of urllib2.urlopen (Python 2 code) not handling TLS
  connections correctly.
