+++ src/utils/error3.py
@@ -1,5 +1,5 @@
 ###
-# Copyright (c) 2015, Ben McGinnes
+# Copyright (c) 2005, Jeremiah Fincher
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -27,9 +27,19 @@
 # POSSIBILITY OF SUCH DAMAGE.
 ###
 
-import sys
+import os
 
-if sys.version_info[0] <= 2:
-    from str2 import *
-else:
-    from str3 import *
+from . import gen
+
+class Error(Exception):
+    def __init__(self, msg, e=None):
+        self.msg = msg
+        self.e = e
+
+    def __str__(self):
+        if self.e is not None:
+            return os.linesep.join([self.msg, gen.exnToString(self.e)])
+        else:
+            return self.msg
+
+# vim:set shiftwidth=4 softtabstop=8 expandtab textwidth=78:
