+++ src/utils/seq.py
@@ -1,5 +1,5 @@
 ###
-# Copyright (c) 2002-2005, Jeremiah Fincher
+# Copyright (c) 2015, Ben McGinnes
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -27,69 +27,9 @@
 # POSSIBILITY OF SUCH DAMAGE.
 ###
 
-def window(L, size):
-    """list * size -> window iterable
+import sys
 
-    Returns a sliding 'window' through the list L of size size."""
-    assert not isinstance(L, int), 'Argument order swapped: window(L, size)'
-    if size < 1:
-        raise ValueError('size <= 0 disallowed.')
-    for i in xrange(len(L) - (size-1)):
-        yield L[i:i+size]
-
-def mapinto(f, L):
-    for (i, x) in enumerate(L):
-        L[i] = f(x)
-
-def renumerate(L):
-    for i in xrange(len(L)-1, -1, -1):
-        yield (i, L[i])
-
-def dameraulevenshtein(seq1, seq2):
-    """Calculate the Damerau-Levenshtein distance between sequences.
-
-    This distance is the number of additions, deletions, substitutions,
-    and transpositions needed to transform the first sequence into the
-    second. Although generally used with strings, any sequences of
-    comparable objects will work.
-
-    Transpositions are exchanges of *consecutive* characters; all other
-    operations are self-explanatory.
-
-    This implementation is O(N*M) time and O(M) space, for N and M the
-    lengths of the two sequences.
-
-    >>> dameraulevenshtein('ba', 'abc')
-    2
-    >>> dameraulevenshtein('fee', 'deed')
-    2
-
-    It works with arbitrary sequences too:
-    >>> dameraulevenshtein('abcd', ['b', 'a', 'c', 'd', 'e'])
-    2
-    """
-    # codesnippet:D0DE4716-B6E6-4161-9219-2903BF8F547F
-    # Conceptually, this is based on a len(seq1) + 1 * len(seq2) + 1 matrix.
-    # However, only the current and two previous rows are needed at once,
-    # so we only store those.
-    # Sourced from http://mwh.geek.nz/2009/04/26/python-damerau-levenshtein-distance/
-    oneago = None
-    thisrow = range(1, len(seq2) + 1) + [0]
-    for x in xrange(len(seq1)):
-        # Python lists wrap around for negative indices, so put the
-        # leftmost column at the *end* of the list. This matches with
-        # the zero-indexed strings and saves extra calculation.
-        twoago, oneago, thisrow = oneago, thisrow, [0] * len(seq2) + [x + 1]
-        for y in xrange(len(seq2)):
-            delcost = oneago[y] + 1
-            addcost = thisrow[y - 1] + 1
-            subcost = oneago[y - 1] + (seq1[x] != seq2[y])
-            thisrow[y] = min(delcost, addcost, subcost)
-            # This block deals with transpositions
-            if (x > 0 and y > 0 and seq1[x] == seq2[y - 1]
-                and seq1[x-1] == seq2[y] and seq1[x] != seq2[y]):
-                thisrow[y] = min(thisrow[y], twoago[y - 2] + 1)
-    return thisrow[len(seq2) - 1]
-
-# vim:set shiftwidth=4 softtabstop=4 expandtab textwidth=79:
-
+if sys.version_info[0] <= 2:
+    from seq2 import *
+else:
+    from seq3 import *
